整体更新backstage服务

+++ myalbum-service-provider-backstage/src/main/java/cn/yan_wm/myalbum/service/provider/backstage/mapper/RolePermissionMapper.java
@@ -17,7 +17,7 @@ public interface RolePermissionMapper extends MyMapper<SysRolePermission> {
      * @param roleId
      * @return
      */
-    List<SysRolePermission> findPermissionIdsByRoleId(@Param("roleId") Long roleId);
+    List<SysRolePermission> findPermissionIdsByRoleId(@Param("roleId") Integer roleId);
 
     /**
      * 通过角色id删除所有的权限
